Move audio tag to top of component
Moved audio tag to top since otherwise it would not always render.
This caused errors sometimes since it is required for the MediaSession API.

Added backup text to audio tag if source not supported.

Added an error handler for when audio doesn't work. It can be called if audio tag doesn't exist, or by the tag itself when there's an error.
Currently tag just prints error on console.